Unison.Hashing.V2

Description

This module exports:

Data types with ContentAddressable instances that correspond to v2 of the Unison hash function.

Miscellaneous helper functions related to hashing.

Documentation

data Branch

Constructors

Branch

Fields

terms :: Map NameSegment (Map Referent MdValues)

types :: Map NameSegment (Map Reference MdValues)

patches :: Map NameSegment Hash

children :: Map NameSegment Hash

data Causal

Constructors

Causal

Fields

branchHash :: Hash

parents :: Set Hash

data DataDeclaration v a

Constructors

DataDeclaration

Fields

modifier :: Modifier

annotation :: a

bound :: [v]

constructors' :: [(a, v, Type v a)]

type Decl v a = Either (EffectDeclaration v a) (DataDeclaration v a)

newtype EffectDeclaration v a

Constructors

EffectDeclaration

Fields

toDataDecl :: DataDeclaration v a

data Kind

Constructors

KindStar

KindArrow Kind Kind

data MatchCase loc a

Constructors

MatchCase (Pattern loc) (Maybe a) a

newtype MdValues

Constructors

MdValues (Set MetadataValue)

data Modifier

Constructors

Structural

Unique Text

newtype NameSegment

A name segment.

Constructors

NameSegment Text

data Patch

Constructors

Patch

Fields

termEdits :: Map Referent (Set TermEdit)

typeEdits :: Map Reference (Set TypeEdit)

data Pattern loc

Constructors

PatternUnbound loc

PatternVar loc

PatternBoolean loc !Bool

PatternInt loc !Int64

PatternNat loc !Word64

PatternFloat loc !Double

PatternText loc !Text

PatternChar loc !Char

PatternConstructor loc !Reference !ConstructorId [Pattern loc]

PatternAs loc (Pattern loc)

PatternEffectPure loc (Pattern loc)

PatternEffectBind loc !Reference !ConstructorId [Pattern loc] (Pattern loc)

PatternSequenceLiteral loc [Pattern loc]

PatternSequenceOp loc (Pattern loc) !SeqOp (Pattern loc)

data Reference

Either a builtin or a user defined (hashed) top-level declaration.

Used for both terms and types. Doesn't distinguish between them.

Other used defined things like local variables don't get References.

Constructors

ReferenceBuiltin Text

ReferenceDerivedId ReferenceId

pattern ReferenceDerived :: Hash -> Pos -> Reference

data ReferenceId

Pos is a position into a cycle of size Size, as cycles are hashed together.

Constructors

ReferenceId Hash Pos

data Referent

Constructors

ReferentRef Reference

ReferentCon Reference ConstructorId

data SeqOp

Constructors

Cons

Snoc

Concat

type Term v a = Term2 v a a v a

Like `Term v`, but with an annotation of type a at every level in the tree

data TermEdit

Constructors

TermEditReplace Referent

TermEditDeprecate

data TermF typeVar typeAnn patternAnn a

Base functor for terms in the Unison language We need typeVar because the term and type variables may differ.

Constructors

TermInt Int64

TermNat Word64

TermFloat Double

TermBoolean Bool

TermText Text

TermChar Char

TermBlank (Blank typeAnn)

TermRef Reference

TermConstructor Reference ConstructorId

TermRequest Reference ConstructorId

TermHandle a a

TermApp a a

TermAnn a (Type typeVar typeAnn)

TermList (Seq a)

TermIf a a a

TermAnd a a

TermOr a a

TermLam a

TermLetRec [a] a

TermLet a a

TermMatch a [MatchCase patternAnn a]

TermTermLink Referent

TermTypeLink Reference

type Type v a = Term TypeF v a

Types are represented as ABTs over the base functor F, with variables in v

data TypeEdit

Constructors

TypeEditReplace Reference

TypeEditDeprecate

data TypeF a

Base functor for types in the Unison language

Constructors

TypeRef Reference

TypeArrow a a

TypeAnn a Kind

TypeApp a a

TypeEffect a a

TypeEffects [a]

TypeForall a

TypeIntroOuter a

hashClosedTerm :: Var v => Term v a -> ReferenceId

hashDecls :: (Eq v, Var v, Show v) => (v -> Name) -> Map v (DataDeclaration v a) -> ResolutionResult a [(v, ReferenceId, DataDeclaration v a)]

compute the hashes of these user defined types and update any free vars corresponding to these decls with the resulting hashes

data List a = Nil | Cons a (List a) becomes something like (List, #xyz, [forall a. #xyz a, forall a. a -> (#xyz a) -> (#xyz a)])

NOTE: technical limitation, this implementation gives diff results if ctors have the same FQN as one of the types. TODO: assert this and bomb if not satisfied, or else do local mangling and unmangling to ensure this doesn't affect the hash.

hashTermComponents :: forall v a extra. Var v => Map v (Term v a, Type v a, extra) -> Map v (ReferenceId, Term v a, Type v a, extra)

hashTermComponentsWithoutTypes :: Var v => Map v (Term v a) -> Map v (ReferenceId, Term v a)

typeToReference :: (Ord v, Show v) => Type v a -> Reference

typeToReferenceMentions :: (Ord v, Show v) => Type v a -> Set Reference

Re-exports

class ContentAddressable a where

A type class that is inhabited by types that can compute a hash of their content.

The base instances of this class should only live in dedicated "hashing packages" such as unison-hashing-v2, whose types and implementations should never change.

Trivial wrapper instances can be written around these, but note these pipelines from MyType ==> SomeHashingType ==> Hash must take care not to change the MyType ==> SomeHashingType conversion, once written.

For example, we might have a representation of some namespace in memory

data Namespace = Namespace Terms Types OtherStuff CachesAndWhatnot

with a somewhat equivalent "hashing" type in some "hashing package", with a ContentAddressable instance

data HashingNamespace = Namespace Terms Types

We can of course make our own convenience instance

instance ContentAddressable Namespace where
  contentHash = contentHash . namespaceToHashingNamespace

But we must make sure that the implementation of namespaceToHashingNamespace never changes the fields in the corresponding HashingNamespace, even as features are added to or removed from Namespace.

Methods

contentHash :: a -> Hash